# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UNIPROM CRM API
  version: 1.0.0
  description: |
    HTTP API CRM для обработки заявок на изготовление изделий.

    Два взаимозаменяемых префикса с одинаковой серверной проверкой прав и области
    видимости: канонический `/api/v1/...` и контрактный `/api/v1/crm/...`.

    Аутентификация — JWT в заголовке `Authorization: Bearer <token>`.
    Технические эндпоинты дополнительно требуют заголовок `x-integration-token`.

    Создание заявки принимает заголовок `Idempotency-Key`. Изменение ресурса
    требует `lockVersion`; при конфликте возвращается 409 `VERSION_CONFLICT`.
  contact:
    name: UNIPROM
servers:
  - url: http://127.0.0.1:3001
    description: Локальный backend
...
paths:
...
  /api/v1/engineering/tasks:
    get:
      tags: [engineering]
      summary: Задания конструкторского отдела
      parameters:
        - name: applicationNumber
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Задания
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EngineeringTask' }
  /api/v1/applications/{number}/engineering-tasks:
    post:
      tags: [engineering]
      summary: Создание задания КО
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                dueAt: { type: string, format: date-time }
      responses:
        '201': { description: Задание создано }
  /api/v1/engineering/tasks/{taskId}/assign:
    post:
      tags: [engineering]
      summary: Назначение конструктора
      description: Назначать конструктора может только роль, которой это разрешено.
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assigneeId]
              properties:
                assigneeId: { type: string }
      responses:
        '200': { description: Назначено }
        '403':
          description: У роли нет права назначать конструктора
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/engineering/tasks/{taskId}/complete:
    post:
      tags: [engineering]
      summary: Завершение задания КО
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: Задание завершено }
  /api/v1/engineering/tasks/{taskId}/cancel:
    post:
      tags: [engineering]
      summary: Отмена задания КО
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200': { description: Задание отменено }
  /api/v1/engineering/conclusions/{conclusionId}/approve:
    post:
      tags: [engineering]
      summary: Утверждение заключения КО
      parameters:
        - name: conclusionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approve, lockVersion]
              properties:
                approve: { type: boolean }
                comment: { type: string }
                lockVersion: { type: integer }
      responses:
        '200': { description: Решение зафиксировано, заявка уведомлена }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/imports/inspect:
    post:
      tags: [imports]
      summary: Разбор загруженного файла
      description: |
        Принимает CSV или XLSX, определяет разделитель и колонки, возвращает
        заголовки и выборку строк. Файл разбирается на сервере, лимит 5000 строк.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Заголовки, выборка и ограничения
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportInspect' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413':
          description: Превышен лимит строк или размер файла
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/imports/upload:
    post:
      tags: [imports]
      summary: Загрузка файла с сопоставлением колонок
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, mapping]
              properties:
                file: { type: string, format: binary }
                mapping:
                  type: string
                  description: JSON-объект вида `{"groupId":"Колонка A", ...}`
      responses:
        '200':
          description: Пакет создан, показан предпросмотр и ошибки
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportPreview' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/imports/validate:
    post:
      tags: [imports]
      summary: Валидация пакета импорта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, mapping, rows]
              properties:
                fileName: { type: string }
                source: { type: string, example: CSV }
                mapping:
                  type: object
                  additionalProperties: { type: string }
                rows:
                  type: array
                  items: { $ref: '#/components/schemas/ImportRow' }
      responses:
        '200':
          description: Пакет в статусе VALIDATED с предпросмотром
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportBatch' }
...
  /api/v1/imports/{batchId}:
    get:
      tags: [imports]
      summary: Пакет импорта с ошибками по строкам
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Пакет и диагностика
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch: { $ref: '#/components/schemas/ImportBatch' }
                  issues:
                    type: array
                    items: { $ref: '#/components/schemas/ImportIssue' }
                  rows:
                    type: array
                    items: { $ref: '#/components/schemas/ImportRow' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    integrationToken:
      type: apiKey
      in: header
      name: x-integration-token
  parameters:
...
    Number:
      name: number
      in: path
      required: true
      schema: { type: string }
      description: Номер заявки.
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string }
...
    BatchId:
      name: batchId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Ошибка запроса
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Конфликт состояния или версии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VERSION_CONFLICT }
        message: { type: string }
        correlationId: { type: string }
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    EngineeringTask:
      type: object
      properties:
        id: { type: string }
        applicationId: { type: string }
        status: { type: string }
        assigneeId: { type: string, nullable: true }
        conclusionDecision: { type: string, nullable: true }
...
    ImportRow:
      type: object
      properties:
        rowNo: { type: integer }
        groupId: { type: string }
        organizationName: { type: string }
        inn: { type: string, nullable: true }
        lineName: { type: string }
        quantity: { type: number }
        unit: { type: string, nullable: true }
        price: { type: number, nullable: true }
    ImportIssue:
      type: object
      properties:
        rowNo: { type: integer }
        field: { type: string }
        reason: { type: string }
        fix: { type: string }
    ImportInspect:
      type: object
      properties:
        fileName: { type: string }
        contentType: { type: string }
        headers:
          type: array
          items: { type: string }
        sample:
          type: array
          items:
            type: object
            additionalProperties: true
        rowCount: { type: integer }
        limit: { type: integer }
    ImportBatch:
      type: object
      properties:
        id: { type: string }
        sourceKey: { type: string }
        fileName: { type: string }
        fileChecksum: { type: string }
        externalBatchId:
          type: string
          nullable: true
          description: Ранее принятый пакет с тем же содержимым, если это повтор.
        status: { type: string, enum: [VALIDATED, COMMITTED, FAILED] }
        totalRows: { type: integer }
        validGroups: { type: integer }
        errorGroups: { type: integer }
        createdGroups: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    ImportPreview:
      type: object
      properties:
        batch: { $ref: '#/components/schemas/ImportBatch' }
        issues:
          type: array
          items: { $ref: '#/components/schemas/ImportIssue' }
        groups:
          type: array
          items:
            type: object
            properties:
              groupId: { type: string }
              rows: { type: integer }
              organizationName: { type: string }
              inn: { type: string, nullable: true }
              totalQty: { type: number }
              totalAmount: { type: number }
              duplicateHint: { type: string, nullable: true }
...
security:
  - bearerAuth: []